Move TriState, FolderDefinition, FolderSettings to Core project
Wire Core→Data and UI→Core project references. Move TriState enum to
Core/Filtering, FolderDefinition and FolderSettings to Core/Folders.
Delete Core/Class1.cs placeholder. Update using statements in UI.
