+++ Core/Filtering/TriState.cs
@@ -1,4 +1,4 @@
-namespace UI.Controls;
+namespace Core.Filtering;
 
 /// <summary>
 /// Three-state filter: Ignored (default), Include (left-click, must have), Exclude (right-click, must not have).
